requirements.txt
fastapi==0.115.11
mock>=4.0.3
pydantic==2.5.0
pytest>=7.0.0
pytest-mock>=3.10.0
python-dotenv==1.0.0
rdflib==7.1.3
requests==2.32.3
ruff==0.9.6
starlette==0.46.0
node-lib @ git+https://${GITHUB_ACCESS_TOKEN}@github.com/National-Digital-Twin/node-lib.git@pre#egg=node-lib
ianode-label-builder @ git+https://${GITHUB_ACCESS_TOKEN}@github.com/National-Digital-Twin/label-builder.git@pre#egg=ianode-label-builder
uvicorn==0.24.0.post1
httpx
sqlalchemy
psycopg2-binary
alembic
geoalchemy2
asyncpg
geoalchemy2[shapely]
